Tank game
This is a game that I wrote in java. It is a one player game, and the user controls a car and attempts to survive for as long as possible. You lose a life if you are hit by an incoming tank, but you have shields. The game also tracks high scores and you have 3 lives until you have to restart. The tanks start coming faster and faster the longer you survive as well.